# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: OTUS Highload Architect
  version: 1.2.0
servers:
- url: /
paths:
...
  /friend/set/{user_id}:
    put:
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: simple
      responses:
        "200":
          description: Пользователь успешно указал своего друга
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-accepts:
      - application/json
  /friend/delete/{user_id}:
    put:
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: simple
      responses:
        "200":
          description: Пользователь успешно удалил из друзей пользователя
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-accepts:
      - application/json
...
  /dialog/{user_id}/send:
    post:
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_dialog__user_id__send_post_request'
      responses:
        "200":
          description: Успешно отправлено сообщение
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-content-type: application/json
      x-accepts:
      - application/json
  /dialog/{user_id}/list:
    get:
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DialogMessage'
                type: array
          description: Диалог между двумя пользователями
        "400":
          description: Невалидные данные ввода
        "401":
          description: Неавторизованный доступ
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_login_post_500_response'
          description: Ошибка сервера
          headers:
            Retry-After:
              description: "Время, через которое еще раз нужно сделать запрос"
              explode: false
              required: false
              schema:
                type: integer
              style: simple
      security:
      - bearerAuth: []
      x-accepts:
      - application/json
components:
...
  schemas:
...
    UserId:
      description: Идентификатор пользователя
      type: string
...
    DialogMessage:
      example:
        from: from
        to: to
        text: "Привет, как дела?"
      properties:
        from:
          description: Идентификатор пользователя
          type: string
        to:
          description: Идентификатор пользователя
          type: string
        text:
          description: Текст сообщения
          example: "Привет, как дела?"
          type: string
      required:
      - from
      - text
      - to
      type: object
...
    _login_post_500_response:
      example:
        code: 0
        message: message
        request_id: request_id
      properties:
        message:
          description: Описание ошибки
          type: string
        request_id:
          description: Идентификатор запроса. Предназначен для более быстрого поиска
            проблем.
          type: string
        code:
          description: Код ошибки. Предназначен для классификации проблем и более
            быстрого решения проблем.
          type: integer
      required:
      - message
      type: object
...
    _dialog__user_id__send_post_request:
      properties:
        text:
          description: Текст сообщения
          example: "Привет, как дела?"
          type: string
      required:
      - text
      type: object
  securitySchemes:
    bearerAuth:
      description: "Авторизация по токену, который был получен в методе /login"
      scheme: bearer
      type: http